Moving usage() func outside the lib
Moving the usage() function outside the file containing the
library functions. The function is moved to the file defining
interfaces for the conversion tool.

Code refactoring.
