+++ tools/adios2pio-nm/adios2pio-nm-lib.h
@@ -7,7 +7,6 @@
 using namespace std;
 
 int ConvertBPToNC(string infilepath, string outfilename, string piotype, MPI_Comm comm_in);
-void usage_nm(string prgname);
 void SetDebugOutput(int val);
 
 #endif /* #ifndef _ADIOS2PIO_NM_LIB_H_ */
