+++ tools/adios2pio-nm/adios2pio-nm.cxx
@@ -2,7 +2,18 @@
 #ifdef TIMING
 #include <gptl.h>
 #endif
+#include <iostream>
 #include "adios2pio-nm-lib.h"
+
+void usage_nm(string prgname)
+{
+    cout << "Usage: " << prgname << " bp_file  nc_file  pio_io_type\n";
+    cout << "   bp file   :  data produced by PIO with ADIOS format\n";
+    cout << "   nc file   :  output file name after conversion\n";
+    cout << "   pio format:  output PIO_IO_TYPE. Supported parameters:\n";
+    cout << "                pnetcdf  netcdf  netcdf4c  netcdf4p   or:\n";
+    cout << "                   1       2        3         4\n";
+}
 
 int main(int argc, char *argv[])
 {
